requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.5
iso8601==1.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pypika-tortoise==0.1.6
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
